requirements.txt
pytest
discord.py
pytest-asyncio
noise
python-dotenv
openai